+++ src/index.ts
@@ -645,3 +645,36 @@ export {
   ToolOutput,
   ToolStatusBadge,
 } from './Tool/Tool';
+
+export {
+  WebPreview,
+  type WebPreviewContextValue,
+  type WebPreviewFactory,
+  type WebPreviewProps,
+} from './WebPreview/WebPreview';
+export {
+  WebPreviewBody,
+  type WebPreviewBodyFactory,
+  type WebPreviewBodyProps,
+} from './WebPreview/WebPreviewBody';
+export {
+  WebPreviewConsole,
+  type WebPreviewConsoleFactory,
+  type WebPreviewConsoleProps,
+  type WebPreviewLog,
+} from './WebPreview/WebPreviewConsole';
+export {
+  WebPreviewNavigation,
+  type WebPreviewNavigationFactory,
+  type WebPreviewNavigationProps,
+} from './WebPreview/WebPreviewNavigation';
+export {
+  WebPreviewNavigationButton,
+  type WebPreviewNavigationButtonFactory,
+  type WebPreviewNavigationButtonProps,
+} from './WebPreview/WebPreviewNavigationButton';
+export {
+  WebPreviewUrl,
+  type WebPreviewUrlFactory,
+  type WebPreviewUrlProps,
+} from './WebPreview/WebPreviewUrl';